requirements.txt
loguru==0.7.2
aiohttp==3.9.3
django_celery_beat==2.6.0
django-ninja==1.1.0
django-easy-audit==1.3.6
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.34.28
botocore==1.34.28
Brotli==1.1.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.1
cssselect2==0.7.0
daphne==4.0.0
dataclasses-json==0.6.4
distro==1.9.0
Django==5.0.1
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-redis==5.4.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
fonttools==4.49.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.5
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.35
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.36
marshmallow==3.21.1
msgpack==1.0.7
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.14.3
orjson==3.10.0
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydyf==0.9.0
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyphen==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.6.0
tinycss2==1.2.1
tqdm==4.66.2
Twisted==23.10.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
yarl==1.9.4
zope.interface==6.1
zopfli==0.2.3
